Handle region-less homepage
If the current location doesn't include a hyphen, it's probably the home
page (at the root or /Home/). Reloading that doesn't work well, as it
just brings up the region picker again. So we check for that and
redirect to en-GB instead. As we've triggered the updating of the region
code already, this will then be picked up by the background script
already and we'll redirect to the proper region-specific homepage
instead